Trying to draw a cube.
Addded a Vertex shader and a pixel shader. Added support in the c++ code
to load these and to create an input layout for the vertex shader.

Added a vertex struct and temporary code in renderer to create vertex and
index buffers for a cube.

Added code for basic forward rendering of a cube, however something is
missing since the cube cannot be seen on screen.

Added .cso to the gitignore file.